refactor: split cmd/glasp/main.go into per-command files
Split the 1803-line main.go into per-command source files (login,
create, clone, pull, push, open, deployment, run_function, convert,
history, config) plus shared helpers (client.go, project.go,
archive.go). Split main_test.go the same way. Pure code movement
within the package; no behavior change.
