+++ cmd/glasp/pull_test.go
@@ -182,3 +182,138 @@ func TestArchivePullRunWritesFailedManifestOnError(t *testing.T) {
 		t.Fatalf("unexpected error: %v", err)
 	}
 }
+func TestPullCommandFlow(t *testing.T) {
+	root := useTempDir(t)
+	if err := config.SaveClaspConfig(root, &config.ClaspConfig{ScriptID: "script-id", RootDir: "src"}); err != nil {
+		t.Fatalf("SaveClaspConfig failed: %v", err)
+	}
+
+	fake := &fakeScriptClient{getContentResp: sampleContent()}
+	orig := newScriptClientWithCacheAuthFn
+	t.Cleanup(func() { newScriptClientWithCacheAuthFn = orig })
+	newScriptClientWithCacheAuthFn = func(ctx context.Context, cachePath, authPath string) (scriptClient, error) {
+		return fake, nil
+	}
+
+	if err := (&PullCmd{}).Run(nil); err != nil {
+		t.Fatalf("PullCmd.Run failed: %v", err)
+	}
+	if _, err := os.Stat(filepath.Join(root, "src", "Code.js")); err != nil {
+		t.Fatalf("expected Code.js to exist: %v", err)
+	}
+	if _, err := os.Stat(filepath.Join(root, "src", "appsscript.json")); err != nil {
+		t.Fatalf("expected appsscript.json to exist: %v", err)
+	}
+}
+
+func TestPullCommandPassesAuthPath(t *testing.T) {
+	root := useTempDir(t)
+	if err := config.SaveClaspConfig(root, &config.ClaspConfig{ScriptID: "script-id", RootDir: "src"}); err != nil {
+		t.Fatalf("SaveClaspConfig failed: %v", err)
+	}
+
+	fake := &fakeScriptClient{getContentResp: sampleContent()}
+	var gotAuthPath string
+	orig := newScriptClientWithCacheAuthFn
+	t.Cleanup(func() { newScriptClientWithCacheAuthFn = orig })
+	newScriptClientWithCacheAuthFn = func(ctx context.Context, cachePath, authPath string) (scriptClient, error) {
+		gotAuthPath = authPath
+		return fake, nil
+	}
+
+	if err := (&PullCmd{Auth: " ./auth/.clasprc.json "}).Run(nil); err != nil {
+		t.Fatalf("PullCmd.Run failed: %v", err)
+	}
+	if gotAuthPath != filepath.Clean("./auth/.clasprc.json") {
+		t.Fatalf("expected cleaned auth path, got %q", gotAuthPath)
+	}
+}
+
+func TestPullCommandRejectsEmptyAuthPath(t *testing.T) {
+	useTempDir(t)
+	orig := newScriptClientWithCacheAuthFn
+	t.Cleanup(func() { newScriptClientWithCacheAuthFn = orig })
+	newScriptClientWithCacheAuthFn = func(ctx context.Context, cachePath, authPath string) (scriptClient, error) {
+		t.Fatal("expected to fail before client creation")
+		return nil, nil
+	}
+
+	err := (&PullCmd{Auth: "   "}).Run(nil)
+	if err == nil {
+		t.Fatalf("expected error for empty --auth path")
+	}
+	if !strings.Contains(err.Error(), "--auth path is empty") {
+		t.Fatalf("unexpected error: %v", err)
+	}
+}
+
+func TestPullCommandDryRunSkipsAPIAndWrites(t *testing.T) {
+	root := useTempDir(t)
+	if err := config.SaveClaspConfig(root, &config.ClaspConfig{ScriptID: "script-id", RootDir: "src"}); err != nil {
+		t.Fatalf("SaveClaspConfig failed: %v", err)
+	}
+
+	orig := newScriptClientWithCacheAuthFn
+	t.Cleanup(func() { newScriptClientWithCacheAuthFn = orig })
+	newScriptClientWithCacheAuthFn = func(ctx context.Context, cachePath, authPath string) (scriptClient, error) {
+		t.Fatal("expected dryrun pull to skip client creation")
+		return nil, nil
+	}
+
+	if err := (&PullCmd{DryRun: true}).Run(nil); err != nil {
+		t.Fatalf("PullCmd.Run failed: %v", err)
+	}
+	if _, err := os.Stat(filepath.Join(root, "src", "Code.js")); !os.IsNotExist(err) {
+		t.Fatalf("expected no pulled files to be written during dryrun")
+	}
+}
+
+func TestPullCommandDryRunRejectsEmptyAuthPath(t *testing.T) {
+	root := useTempDir(t)
+	if err := config.SaveClaspConfig(root, &config.ClaspConfig{ScriptID: "script-id", RootDir: "src"}); err != nil {
+		t.Fatalf("SaveClaspConfig failed: %v", err)
+	}
+	orig := newScriptClientWithCacheAuthFn
+	t.Cleanup(func() { newScriptClientWithCacheAuthFn = orig })
+	newScriptClientWithCacheAuthFn = func(ctx context.Context, cachePath, authPath string) (scriptClient, error) {
+		t.Fatal("expected to fail before client creation")
+		return nil, nil
+	}
+
+	err := (&PullCmd{DryRun: true, Auth: "   "}).Run(nil)
+	if err == nil {
+		t.Fatalf("expected error for empty --auth path in dryrun pull")
+	}
+	if !strings.Contains(err.Error(), "--auth path is empty") {
+		t.Fatalf("unexpected error: %v", err)
+	}
+}
+
+func TestPullCommandRejectsTSXFileExtension(t *testing.T) {
+	root := useTempDir(t)
+	cfg := &config.ClaspConfig{
+		ScriptID: "script-id",
+		RootDir:  "src",
+		Extra: map[string]json.RawMessage{
+			"fileExtension": json.RawMessage(`"tsx"`),
+		},
+	}
+	if err := config.SaveClaspConfig(root, cfg); err != nil {
+		t.Fatalf("SaveClaspConfig failed: %v", err)
+	}
+
+	orig := newScriptClientWithCacheAuthFn
+	t.Cleanup(func() { newScriptClientWithCacheAuthFn = orig })
+	newScriptClientWithCacheAuthFn = func(ctx context.Context, cachePath, authPath string) (scriptClient, error) {
+		t.Fatal("expected to fail before client creation")
+		return nil, nil
+	}
+
+	err := (&PullCmd{}).Run(nil)
+	if err == nil {
+		t.Fatalf("expected error for tsx fileExtension")
+	}
+	if !strings.Contains(err.Error(), `fileExtension "tsx" is not supported`) {
+		t.Fatalf("unexpected error: %v", err)
+	}
+}
